Refresh/requirements.txt
google-cloud-firestore
google-cloud-secret-manager
tenacity
firebase-admin==6.0.1
requests==2.28.1
python-dotenv==0.21.0
Flask==2.3.2
Flask-Cors==3.0.10
pytz==2023.3
google-cloud-functions==1.0.0
